Added logging logic in admin Controller. Fixed bug with logger.

+++ nwcserver/src/com/nwchecker/server/controller/AdminOptionsController.java
@@ -3,11 +3,13 @@ package com.nwchecker.server.controller;
 import java.security.Principal;
 import java.util.List;
 
-import javax.servlet.http.HttpServletRequest;
-
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -30,6 +32,8 @@ import com.nwchecker.server.service.UserService;
 @Controller
 public class AdminOptionsController {
 
+	private static final Logger LOG = Logger.getLogger(ContestController.class);
+	
 	@Autowired
 	private UserService userService;
 	
@@ -43,53 +47,91 @@ public class AdminOptionsController {
 		dataBinder.setValidator(validator);
 	}
 	
-	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	@RequestMapping(value = "/admin", method = RequestMethod.GET)
-	public String admin() {
-		return "adminOptions/users";
+	public String admin(Principal principal) {
+		UserDetails currentUser = 
+				(UserDetails) ((Authentication) principal).getPrincipal();
+		LOG.info(currentUser.getUsername() + " try to access administration page");
+		GrantedAuthority admin = new SimpleGrantedAuthority("ROLE_ADMIN");
+		if (currentUser.getAuthorities().contains(admin)) {
+			LOG.info(currentUser.getUsername() + " -Access GRANTED-");
+			return "adminOptions/users";
+		}
+		LOG.warn("Only administrators can access administration pages!");
+		LOG.info(currentUser.getUsername() + " -Access DENIED-");
+		return "redirect: /index.do";
 	}
 	
-	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	@RequestMapping(value = "/getUsers", method = RequestMethod.GET)
-	public @ResponseBody String getUsers() {
-		List<User> users = userService.getUsers();
-		Gson gson = 
-				new GsonBuilder().excludeFieldsWithoutExposeAnnotation().create();
-		return gson.toJson(users.toArray());
+	public @ResponseBody String getUsers(Principal principal) {
+		UserDetails currentUser = 
+				(UserDetails) ((Authentication) principal).getPrincipal();
+		LOG.info(currentUser.getUsername() + " try to receive users list");
+		GrantedAuthority admin = new SimpleGrantedAuthority("ROLE_ADMIN");
+		if (currentUser.getAuthorities().contains(admin)) {
+			LOG.info(currentUser.getUsername() + " -Access GRANTED-");
+			LOG.info("Users list received by " + currentUser.getUsername());
+			List<User> users = userService.getUsers();
+			Gson gson = 
+					new GsonBuilder().excludeFieldsWithoutExposeAnnotation().create();
+			return gson.toJson(users.toArray());
+		}
+		LOG.warn("Only administrators can access users list!");
+		LOG.info(currentUser.getUsername() + " -Access DENIED-");
+		return "You haven't rights for this operation!";
 	}
 	
-	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	@RequestMapping(value = "/userEdit", method = RequestMethod.GET)
-	public String user(HttpServletRequest request, Model model) {
-		String username = request.getParameter("Username");
-		if (username == null) {
+	public String user(@RequestParam("Username") String username, 
+						Model model, Principal principal) {
+		UserDetails currentUser = 
+				(UserDetails) ((Authentication) principal).getPrincipal();
+		LOG.info(currentUser.getUsername() + " try to acces user edit page");
+		GrantedAuthority admin = new SimpleGrantedAuthority("ROLE_ADMIN");
+		if (currentUser.getAuthorities().contains(admin)) {
+			LOG.info(currentUser.getUsername() + " -Access GRANTED-");
+			User user = userService.getUserByUsername(username);
+			user.setPassword("");
+			model.addAttribute("userData", user);
+			return "adminOptions/userEdit";
+		}
+		LOG.warn("Only administrators can access administration pages!");
+		LOG.info(currentUser.getUsername() + " -Access DENIED-");
+		return "redirect: /index.do";
+	}
+	
+	@RequestMapping(value = "/changeUser", method = RequestMethod.POST)
+	public String changeUser(@ModelAttribute("userData") @Validated User userData, 
+							BindingResult result, Principal principal) {
+		UserDetails currentUser = 
+				(UserDetails) ((Authentication) principal).getPrincipal();
+		LOG.info(currentUser.getUsername() + " try to change user " + userData.getUsername());
+		GrantedAuthority admin = new SimpleGrantedAuthority("ROLE_ADMIN");
+		if (currentUser.getAuthorities().contains(admin)) {
+			LOG.info(currentUser.getUsername() + " -Access GRANTED-");
+			LOG.info("User can be changed");
+			if (result.hasErrors()) {
+				LOG.warn("Verification failed!");
+				return "/adminOptions/userEdit";
+			}
+			User user = userService.getUserByUsername(userData.getUsername());
+			if (!user.getRoles().isEmpty()) {
+				user.setRoles(null);
+				userService.deleteUserRoles(user);
+			}
+			user = setUserPassword(user, userData.getPassword());
+			user.setDisplayName(userData.getDisplayName());
+			user.setEmail(userData.getEmail());
+			user = setUserRoles(user, userData.getRolesDesc());
+			user.setDepartment(userData.getDepartment());
+			user.setInfo(userData.getInfo());		
+			userService.updateUser(user);
+			LOG.info("User " + user.getUsername() + " changed by " + currentUser.getUsername());
 			return "redirect:admin.do";
 		}
-		User user = userService.getUserByUsername(username);
-		user.setPassword("");
-		model.addAttribute("userData", user);
-		return "adminOptions/userEdit";
-	}
-	
-	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	@RequestMapping(value = "/changeUser", method = RequestMethod.POST)
-	public String changeUser(@ModelAttribute("userData") @Validated User userData, BindingResult result) {
-		if (result.hasErrors()) {
-			return "/adminOptions/userEdit";
-		}
-		User user = userService.getUserByUsername(userData.getUsername());
-		if (!user.getRoles().isEmpty()) {
-			user.setRoles(null);
-			userService.deleteUserRoles(user);
-		}
-		user = setUserPassword(user, userData.getPassword());
-		user.setDisplayName(userData.getDisplayName());
-		user.setEmail(userData.getEmail());
-		user = setUserRoles(user, userData.getRolesDesc());
-		user.setDepartment(userData.getDepartment());
-		user.setInfo(userData.getInfo());		
-		userService.updateUser(user);
-		return "redirect:admin.do";
+		LOG.warn("Only administrators can change users data!");
+		LOG.info(currentUser.getUsername() + " -Access DENIED-");
+		return "redirect: /index.do";
 	}
 	
 	private User setUserPassword(User user, String password) {
@@ -118,12 +160,25 @@ public class AdminOptionsController {
 		return user;
 	}
 	
-	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	@RequestMapping(value = "/deleteUser", method = RequestMethod.POST)
 	public String deleteUser(@RequestParam("username") String username, Principal principal) {
-		if (!principal.getName().equals(username)) {
-			userService.deleteUserByName(username);
+		UserDetails currentUser = 
+				(UserDetails) ((Authentication) principal).getPrincipal();
+		LOG.info(currentUser.getUsername() + " try to delete user " + username);
+		GrantedAuthority admin = new SimpleGrantedAuthority("ROLE_ADMIN");
+		if (currentUser.getAuthorities().contains(admin)) {
+			LOG.info(currentUser.getUsername() + " -Access GRANTED-");
+			LOG.info("User can be deleted");
+			if (!principal.getName().equals(username)) {
+				userService.deleteUserByName(username);
+				LOG.info("User " + username + " deleted by " + currentUser.getUsername());
+			} else {
+				LOG.warn("User cann't delete themself!");
+			}
+			return "redirect:admin.do";
 		}
-		return "redirect:admin.do";
+		LOG.warn("Only administrators can change users data!");
+		LOG.info(currentUser.getUsername() + " -Access DENIED-");
+		return "redirect: /index.do";
 	}
 }
